Refactoring. Checking qrs multilead
