+++ Source/Model/main/delineation/p/gamma/xtd.py
