+++ Source/Model/main/params/qrs.py
@@ -50,10 +50,7 @@ QRSParams['GAMMA_LEFT_ORIGIN_INCORRECT'] = QRSDefaultParams.GAMMA_LEFT_ORIGIN_IN
 QRSParams['GAMMA_RIGHT_ORIGIN_INCORRECT'] = QRSDefaultParams.GAMMA_RIGHT_ORIGIN_INCORRECT
 QRSParams['GAMMA_BORD_SCALE'] = QRSDefaultParams.GAMMA_BORD_SCALE
 
-QRSParams['DELTA_MEAN_QRS_DIFF_CORR'] = QRSDefaultParams.DELTA_MEAN_QRS_DIFF_CORR
 QRSParams['DELTA_MEAN_QRS_LOC'] = QRSDefaultParams.DELTA_MEAN_QRS_LOC
 QRSParams['DELTA_MIN_QRS_FOUND'] = QRSDefaultParams.DELTA_MIN_QRS_FOUND
 QRSParams['DELTA_MAX_QRS_LOST'] = QRSDefaultParams.DELTA_MAX_QRS_LOST
 
-QRSParams['EXTRA_BEAT_PART'] = QRSDefaultParams.EXTRA_BEAT_PART
-QRSParams['MUTABILITY_PERCENTAGE'] = QRSDefaultParams.MUTABILITY_PERCENTAGE
