+++ Source/Model/main/params/qrs_default.py
@@ -48,13 +48,9 @@ class QRSDefaultParams:
     GAMMA_RIGHT_ORIGIN_INCORRECT = 0.005
     GAMMA_BORD_SCALE = 2
 
-    DELTA_MEAN_QRS_DIFF_CORR = 0.25
     DELTA_MEAN_QRS_LOC = 0.75
     DELTA_MIN_QRS_FOUND = 0.5
     DELTA_MAX_QRS_LOST = 0.3
-
-    EXTRA_BEAT_PART = 0.7
-    MUTABILITY_PERCENTAGE = 0.3
